add post endpoint for creating conversations
