+++ backend/src/chat/Message.js
@@ -1,0 +1,9 @@
+class Message {
+    constructor(sender, message) {
+        this.sender = sender;
+        this.message = message;
+        this.timestamp = Date.now();
+    }
+}
+
+module.exports = Message;
