+++ backend/src/routes/chat.js
@@ -9,7 +9,6 @@ router.get('/conversations/:conversationId', async (req, res, next) => {
         res.json(messages);
         next();
     } catch (error) {
-        console.error(error);
         res.status(404).json({ error: error });
         next(error);
     }
@@ -21,7 +20,19 @@ router.get('/conversations/user/:userId', async (req, res, next) => {
         res.json(conversations);
         next();
     } catch (error) {
-        console.error(error);
+        res.status(404).json({ error: error });
+        next(error);
+    }
+});
+
+router.post('/conversations/user/:userId', async (req, res, next) => {
+    try {
+        const receivingUserId = req.query.to;
+        const sendingUserId = req.params.userId;
+        const conversation = await messageStore.sendMessage(req.body.content, sendingUserId, receivingUserId);
+        res.status(201).json(conversation);
+        next();
+    } catch (error) {
         res.status(404).json({ error: error });
         next(error);
     }
